fix(main/libwebp): add regex to filter rc releases for auto update
Earlier the package was updated to 1.6.0-rc1. Hopefully,
https://github.com/webmproject/libwebp/compare/v1.6.0-rc1...v1.6.0 shows
that there is no change at all in the builds of both the versions.
Avoiding adding +really hack to change the version as it will break
upcoming auto updates. And adding a epoch is unnecessary, we need to
ensure that we don't use epochs unnecessarily.

This should automatically resolve the issue when the next auto-update
happens. I'll ignore the errors for libwebp in repository health checks
for the time being

%ci:no-build